garlnd/requirements.txt
aiohttp==3.8.1
channels==3.0.4
daphne==3.0.2
Django==4.0
django-bootstrap5==21.3
django-brake==1.6.0
django-ranged-response==0.2.0
django-registration-redux==2.9
django-simple-captcha==0.5.14
Pillow==9.0.0
psycopg2-binary==2.9.3
pytz